+++ truffle/CHANGELOG.md
@@ -4,7 +4,8 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 
 ## Version 20.0.0 Beta 2
 * `@GenerateUncached` is now inherited by subclasses.
-* `NodeFactory` now supports `getUncachedInstance` that returns the uncached singleton.  
+* `NodeFactory` now supports `getUncachedInstance` that returns the uncached singleton. 
+* `@GenerateUncached` can now be used in combination with `@NodeChild` if execute signatures for all arguments are present.
 
 ## Version 20.0.0 Beta 1
 * The [LanguageProvider](https://www.graalvm.org/truffle/javadoc/org/graalvm/polyglot/tck/LanguageProvider.html#createIdentityFunctionSnippet-org.graalvm.polyglot.Context-) can override the default verfication of the TCK `IdentityFunctionTest`.